+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Reduced default max speed in `WalkingRobotIKEnv` (from 15 to 8)
 - Added velocity in `WalkingRobotIKEnv` observation
 - Added `SE_SERVER_ADDR` env variable
+- Added `symmetric_control` and `allowed_leg_angle` to limit action space
+- Longer episodes and limited at 10Hz
 
 # Release 0.0.1